chore: move LibEIP155TxStruct to core-utils
