+++ packages/core-utils/src/coders/index.ts
@@ -1,2 +1,3 @@
 export * from './types'
 export * from './sequencer-batch'
+export * from './accounts'
